scd: moved ssl-utilities to libcommon
The utilities are reqired in the SCD and the remote
attestation tool and therefore moved to the common
library.
