+++ common/ssl_util.c
